Split fixtures in the Site module
