+++ saleor/site/tests/fixtures/__init__.py
@@ -1,0 +1,1 @@
+from .site_settings import *  # noqa: F401, F403
